test: add id3 reader test

+++ app/api/common/blob/blobUtil.ts
@@ -58,7 +58,7 @@ export default class BlobUtil {
     const dataView: DataView = new DataView(buffer);
     let offset = 0;
     dataViews.map((val: DataView) => val.buffer).forEach((val: ArrayBuffer) => {
-      typedArr.set(new Uint16Array(val), offset);
+      typedArr.set(new Uint8Array(val), offset);
       offset += val.byteLength;
     });
     return dataView;
